+++ config/apt/20_proxmox_1_apt.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 if grep -q Ubuntu /proc/version; then
-    ( ( sleep 1 && rm -- "${0}" ) & )
+    [ -f "${0}" ] && rm -- "${0}"
     exit 0
 fi
 
@@ -140,4 +140,4 @@ ufw status verbose
 sync
 
 # cleanup
-rm -- "${0}"
+[ -f "${0}" ] && rm -- "${0}"
