+++ config/pacman/90_archiso_build.sh
@@ -78,4 +78,4 @@ popd
 sync
 
 # cleanup
-rm -- "${0}"
+[ -f "${0}" ] && rm -- "${0}"
